# Make varint serialization faster