--- a/source/postcard/src/varint.rs
+++ b/source/postcard/src/varint.rs
@@ -25,79 +25,64 @@ pub const fn max_of_last_byte<T: Sized>() -> u8 {
 #[inline]
 pub fn varint_usize(n: usize, out: &mut [u8; varint_max::<usize>()]) -> &mut [u8] {
     let mut value = n;
-    for i in 0..varint_max::<usize>() {
-        out[i] = value.to_le_bytes()[0];
-        if value < 128 {
-            return &mut out[..=i];
-        }
-
-        out[i] |= 0x80;
+    let mut i = 0;
+    while value >= 128 {
+        out[i] = value.to_le_bytes()[0] | 0b10000000;
         value >>= 7;
+        i += 1;
     }
-    debug_assert_eq!(value, 0);
-    &mut out[..]
+    out[i] = value as u8;
+    &mut out[..=i]
 }
 
 #[inline]
 pub fn varint_u16(n: u16, out: &mut [u8; varint_max::<u16>()]) -> &mut [u8] {
     let mut value = n;
-    for i in 0..varint_max::<u16>() {
-        out[i] = value.to_le_bytes()[0];
-        if value < 128 {
-            return &mut out[..=i];
-        }
-
-        out[i] |= 0x80;
+    let mut i = 0;
+    while value >= 128 {
+        out[i] = value.to_le_bytes()[0] | 0b10000000;
         value >>= 7;
+        i += 1;
     }
-    debug_assert_eq!(value, 0);
-    &mut out[..]
+    out[i] = value as u8;
+    &mut out[..=i]
 }
 
 #[inline]
 pub fn varint_u32(n: u32, out: &mut [u8; varint_max::<u32>()]) -> &mut [u8] {
     let mut value = n;
-    for i in 0..varint_max::<u32>() {
-        out[i] = value.to_le_bytes()[0];
-        if value < 128 {
-            return &mut out[..=i];
-        }
-
-        out[i] |= 0x80;
+    let mut i = 0;
+    while value >= 128 {
+        out[i] = value.to_le_bytes()[0] | 0b10000000;
         value >>= 7;
+        i += 1;
     }
-    debug_assert_eq!(value, 0);
-    &mut out[..]
+    out[i] = value as u8;
+    &mut out[..=i]
 }
 
 #[inline]
 pub fn varint_u64(n: u64, out: &mut [u8; varint_max::<u64>()]) -> &mut [u8] {
     let mut value = n;
-    for i in 0..varint_max::<u64>() {
-        out[i] = value.to_le_bytes()[0];
-        if value < 128 {
-            return &mut out[..=i];
-        }
-
-        out[i] |= 0x80;
+    let mut i = 0;
+    while value >= 128 {
+        out[i] = value.to_le_bytes()[0] | 0b10000000;
         value >>= 7;
+        i += 1;
     }
-    debug_assert_eq!(value, 0);
-    &mut out[..]
+    out[i] = value as u8;
+    &mut out[..=i]
 }
 
 #[inline]
 pub fn varint_u128(n: u128, out: &mut [u8; varint_max::<u128>()]) -> &mut [u8] {
     let mut value = n;
-    for i in 0..varint_max::<u128>() {
-        out[i] = value.to_le_bytes()[0];
-        if value < 128 {
-            return &mut out[..=i];
-        }
-
-        out[i] |= 0x80;
+    let mut i = 0;
+    while value >= 128 {
+        out[i] = value.to_le_bytes()[0] | 0b10000000;
         value >>= 7;
+        i += 1;
     }
-    debug_assert_eq!(value, 0);
-    &mut out[..]
+    out[i] = value as u8;
+    &mut out[..=i]
 }