@jamesmunns 2025-09-14
Relevant past changes: #60, #62.

@ldm0 would you mind fixing the CI failures for the new benchmark stuff? For the existing code you didn't change, you can `#[allow(...)]` for now and I'll clean those up later.